Move session to models/login
